+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 
 # Unreleased
 
+<<<<<<< main
 - Export `core/account` folder and the functions: `createObjectAddress` `createResourceAddress` `createTokenAddress`
+=======
+- [`Fix`] Respect pagination arguments on `Events` queries
+>>>>>>> main
 
 # 1.20.0 (2024-06-18)